preprocess/requirements.txt
Pillow
pytest
pytest-pep8
numpy
matplotlib